remove keyboard class
